+++ src/components/Icons/CheckboxIcon.tsx
