+++ src/components/Icons/ChervonDoubleIcon.tsx
