+++ src/components/Icons/ChervonUpIcon.tsx
