+++ src/components/Icons/ConfigurationIcon.tsx
