+++ src/components/Icons/FileExportIcon.tsx
