+++ src/components/Icons/RemoveIcon.tsx
