+++ src/components/Icons/SVGWrapper/createSVGWrapper.tsx
