+++ src/components/Icons/SellsIcon.tsx
