+++ src/components/Icons/StorefrontIcon.tsx
