+++ src/components/Icons/TranslationsIcon.tsx
