+++ src/components/Icons/VouchersIcon.tsx
